move src to commonMain about app
