+++ app/src/androidMain/kotlin/jp/co/yumemi/droidtraining/MainActivity.kt
@@ -9,7 +9,6 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.ui.Modifier
-import jp.co.yumemi.droidtraining.components.YumemiApp
 import jp.co.yumemi.droidtraining.core.model.ThemeConfig
 import jp.co.yumemi.droidtraining.core.ui.shouldUseDarkTheme
