+++ app/src/commonMain/kotlin/jp/co/yumemi/droidtraining/di/Koin.kt
@@ -5,7 +5,6 @@ import jp.co.yumemi.droidtraining.core.repository.di.RepositoryModule
 import jp.co.yumemi.droidtraining.feature.detail.di.DetailModule
 import jp.co.yumemi.droidtraining.feature.home.di.HomeModule
 import org.koin.core.KoinApplication
-import org.koin.ksp.generated.module
 
 fun KoinApplication.applyModules() {
     // App
